+++ tests/ClientTest.php
@@ -12,6 +12,7 @@ use Hampel\XenForo\Api\Resource\Attachments;
 use Hampel\XenForo\Api\Resource\Auth;
 use Hampel\XenForo\Api\Resource\ConversationMessages;
 use Hampel\XenForo\Api\Resource\Conversations;
+use Hampel\XenForo\Api\Resource\Featured;
 use Hampel\XenForo\Api\Resource\Forums;
 use Hampel\XenForo\Api\Resource\Index;
 use Hampel\XenForo\Api\Resource\Me;
@@ -21,6 +22,7 @@ use Hampel\XenForo\Api\Resource\MediaCategories;
 use Hampel\XenForo\Api\Resource\MediaComments;
 use Hampel\XenForo\Api\Resource\Nodes;
 use Hampel\XenForo\Api\Resource\OAuth2;
+use Hampel\XenForo\Api\Resource\OEmbed;
 use Hampel\XenForo\Api\Resource\Posts;
 use Hampel\XenForo\Api\Resource\ProfilePostComments;
 use Hampel\XenForo\Api\Resource\ProfilePosts;
@@ -30,6 +32,8 @@ use Hampel\XenForo\Api\Resource\ResourceReviews;
 use Hampel\XenForo\Api\Resource\ResourceUpdates;
 use Hampel\XenForo\Api\Resource\ResourceVersions;
 use Hampel\XenForo\Api\Resource\Search;
+use Hampel\XenForo\Api\Resource\SearchForums;
+use Hampel\XenForo\Api\Resource\Stats;
 use Hampel\XenForo\Api\Resource\Threads;
 use Hampel\XenForo\Api\Resource\Users;
 use PHPUnit\Framework\Attributes\DataProvider;
@@ -110,6 +114,10 @@ final class ClientTest extends TestCase
             'profilePosts' => ProfilePosts::class,
             'profilePostComments' => ProfilePostComments::class,
             'search' => Search::class,
+            'searchForums' => SearchForums::class,
+            'featured' => Featured::class,
+            'stats' => Stats::class,
+            'oembed' => OEmbed::class,
             'attachments' => Attachments::class,
             'media' => Media::class,
             'mediaAlbums' => MediaAlbums::class,